Implement group management
Group creation is automatic so long as members are added to it and
deletion is handled in the same API request as the membership set
call by setting the group to be empty and submitting a boolean for
the isRemove JSON field of the GroupParameters structure.

Implement the API calls using serde for struct marshalling to camel
case from snake case, porting call target and structure from the Go
code.

Testing:
  Verified creation of group with test-user added to it
  Verified removal of group once emptied of users